Unify all add/remove/save methods
With validation and business rules abtracted out, it's easy to have
common implementations for add/remove/save. See
doAdd(),doSave(),doRemove()
